fb: mm: fix page_alloc_stats and pageblock stats
Calling this_cpu_ptr() from a preeemptible code raises a warning
if CONFIG_PREEMPT_DEBUG is on.

Percpu data was used to track page allocation requests and failures.
This data is used for the long-term monitoring/tracking, so we're fine
with semi-accurate data and really want to minimize the overhead.
This is why this_cpu_ptr() was used, which is fine while CONFIG_PREEMPT_DEBUG
is off. If CONFIG_PREEMPT_DEBUG is enabled it raises bogus warnings
and can cause stalls. Example trace:

[  608.815659] BUG: using smp_processor_id() in preemptible [00000000] code: swapper/0/1
[  608.815967] caller is __alloc_pages_nodemask+0xac/0x620
[  608.816967] CPU: 0 PID: 1 Comm: swapper/0 Not tainted 5.12.0-rc5-00059-g1ae9c61458da #1
[  608.817954] Hardware name: Quanta Twin Lakes MP/Twin Lakes Passive MP, BIOS F09_3A21 05/25/2020
[  608.817954] Call Trace:
[  608.817954]  dump_stack+0xa4/0xe5
[  608.817954]  check_preemption_disabled+0xb6/0xd0
[  608.817954]  __alloc_pages_nodemask+0xac/0x620
[  608.817954]  ? __alloc_pages_slowpath.constprop.98+0x1450/0x1450
[  608.817954]  ? _find_next_bit.constprop.1+0x3e/0xf0
[  608.817954]  ? find_first_bit+0x19/0x70
[  608.817954]  alloc_page_interleave+0x12/0xf0
[  608.817954]  allocate_slab+0x30a/0x4c0
[  608.817954]  ___slab_alloc+0x4a1/0x7e0
[  608.817954]  ? create_object+0x3f/0x550
[  608.817954]  ? __slab_alloc.isra.76+0x7b/0x100
[  608.817954]  ? create_object+0x3f/0x550
[  608.817954]  ? __slab_alloc.isra.76+0xa7/0x100
[  608.817954]  __slab_alloc.isra.76+0xa7/0x100
[  608.817954]  ? create_object+0x3f/0x550
[  608.817954]  ? create_object+0x3f/0x550
[  608.817954]  kmem_cache_alloc+0x25e/0x2d0
[  608.817954]  create_object+0x3f/0x550
[  608.817954]  ? kasan_unpoison+0x38/0x60
[  608.817954]  slab_post_alloc_hook+0x6e/0x560
[  608.817954]  ? lockdep_hardirqs_on+0x79/0x100
[  608.817954]  __kmalloc+0x15a/0x250
[  608.817954]  acpi_ns_internalize_name+0x98/0x106

Fix this by using raw_cpu_ptr() instead.

Also, a similar problem exists around pageblock statistics. But there we want
the accuracy (and it's not the hot path), so let's simple disable the preemption
around changing the pageblock stats.

Signed-off-by: Roman Gushchin <guro@fb.com>