fb: percpu: percpu depopulation hotfix
If a percpu chunk is sidelined and the last outstanding allocation
is freed, we're missing the "else if" clause which is responsible
for putting the chunk into the depopualtion queue and eventually
get rid of it. The problem was reported upstream on powerpc
(because of bad luck it doesn't reveal itself on x86 during th test).

Fix this by explicitly clearing the depopulated flag and moving
the fully empty chunk into the free chunks slot.

In upstgream the fix will be merged into the original patch with
some renamings.

Signed-off-by: Roman Gushchin <guro@fb.com>