FB-ONLY: make the unlink fsnotify events happen after d_delete
Summary:
FB-ONLY: make the unlink fsnotify events happen after d_delete

Watchman reported a problem where their monitoring tool would see a file
unlink, but would immediately stat that file and see that it still
existed.  This was caused by the fsnotify people moving the notification
to right after the unlink, meaning there's a gap in time where the file
is unlinked but the dentry still exists.

This will be fixed upstream by a rather large rework, but the actual fix
is simple enough.  Simply remove the previous notifications and bundle
it with a d_delete helper that will hold onto the dentry->d_inode until
after we've emitted the notification.  This fixes the reproducer that I
wrote.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>