btrfs: add wrapper for conditional start of exclusive operation
Summary:
btrfs: add wrapper for conditional start of exclusive operation

To support optional cancellation of some operations, add helper that will
wrap all the combinations. In normal mode it's same as
btrfs_exclop_start, in cancellation mode it checks if it's already
running and request cancellation and waits until completion.

The error codes can be returned to to user space and semantics is not
changed, adding ECANCELED. This should be evaluated as an error and that
the operation has not completed and the operation should be restarted
or the filesystem status reviewed.


(cherry picked from commit 17aaa434ed39cbad48824ef4bb9ec3707091ae5b)

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: David Sterba <dsterba@suse.com>