kdump: Disable more stuff. Reenable kallsyms
Summary: I ran 'nm --size-sort  vmlinux' and disabled a bunch more stuff
that was just taking up a sizable amount of space.

Also turn KALLSYMS back on because it's useful when shit goes sideways,
and disable HUNGTASK once again, because when we're in writeback with
one CPU, it fires, and eventually reboots the machine.

Test Plan: tested on kerneltest host

Reviewers: #kernel_ops

Subscribers:

Tasks:

Tags: